backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fotoerp-backend"
version = "0.1.0"
description = "FotoDerp Backend — Bildanalyse, Metadaten, KI-Pipeline"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "pillow>=11.0",
    "exifread>=3.0",
    "pydantic>=2.10",
]

[tool.ruff]
target-version = "py311"
line-length = 100
